# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MongoVision API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /connections/{connectionId}/databases/{dbName}/collections:
    get:
      operationId: listCollections
      tags: [collections]
      summary: List all collections in a database
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of collections
          content:
            application/json:
              schema:
                type: object
                properties:
                  collections:
                    type: array
                    items:
                      $ref: "#/components/schemas/Collection"
                required: [collections]
    post:
      operationId: createCollection
      tags: [collections]
      summary: Create a new collection
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCollectionBody"
      responses:
        "200":
          description: Collection created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /connections/{connectionId}/databases/{dbName}/collections/{collectionName}/query:
    post:
      operationId: executeQuery
      tags: [query]
      summary: Execute a query against a collection
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
        - name: collectionName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryBody"
      responses:
        "200":
          description: Query results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResult"
  /connections/{connectionId}/databases/{dbName}/collections/{collectionName}/aggregate:
    post:
      operationId: executeAggregate
      tags: [query]
      summary: Execute an aggregation pipeline
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
        - name: collectionName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AggregateBody"
      responses:
        "200":
          description: Aggregation results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResult"
...
  /connections/{connectionId}/databases/{dbName}/collections/{collectionName}/indexes:
    get:
      operationId: listIndexes
      tags: [indexes]
      summary: List all indexes for a collection
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
        - name: collectionName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of indexes
          content:
            application/json:
              schema:
                type: object
                properties:
                  indexes:
                    type: array
                    items:
                      $ref: "#/components/schemas/IndexInfo"
                required: [indexes]
    post:
      operationId: createIndex
      tags: [indexes]
      summary: Create an index
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
        - name: collectionName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIndexBody"
      responses:
        "200":
          description: Index created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /connections/{connectionId}/databases/{dbName}/collections/{collectionName}/explain:
    post:
      operationId: explainQuery
      tags: [performance]
      summary: Explain a query execution plan
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
        - name: collectionName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryBody"
      responses:
        "200":
          description: Query explain plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExplainResult"
  /connections/{connectionId}/databases/{dbName}/collections/{collectionName}/index-suggestions:
    post:
      operationId: suggestIndexes
      tags: [performance]
      summary: Suggest indexes based on a query
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
        - name: collectionName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryBody"
      responses:
        "200":
          description: Index suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndexSuggestions"
  /connections/{connectionId}/databases/{dbName}/collections/{collectionName}/export:
    post:
      operationId: exportCollection
      tags: [import-export]
      summary: Export collection data
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
        - name: collectionName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportBody"
      responses:
        "200":
          description: Exported data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResult"
  /connections/{connectionId}/databases/{dbName}/collections/{collectionName}/import:
    post:
      operationId: importCollection
      tags: [import-export]
      summary: Import data into a collection
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
        - name: collectionName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Collection:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        documentCount:
          type: number
        storageSize:
          type: number
        totalIndexSize:
          type: number
        avgDocumentSize:
          type: number
      required: [name, type]
    CreateCollectionBody:
      type: object
      properties:
        name:
          type: string
      required: [name]
...
    QueryBody:
      type: object
      properties:
        filter:
          type: object
          additionalProperties: true
        sort:
          type: object
          additionalProperties: true
        projection:
          type: object
          additionalProperties: true
        limit:
          type: number
        skip:
          type: number
      required: []
    AggregateBody:
      type: object
      properties:
        pipeline:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [pipeline]
    QueryResult:
      type: object
      properties:
        documents:
          type: array
          items:
            type: object
            additionalProperties: true
        count:
          type: number
        executionTimeMs:
          type: number
      required: [documents, count, executionTimeMs]
...
    IndexInfo:
      type: object
      properties:
        name:
          type: string
        key:
          type: object
          additionalProperties: true
        unique:
          type: boolean
        sparse:
          type: boolean
        background:
          type: boolean
        size:
          type: number
      required: [name, key]
    CreateIndexBody:
      type: object
      properties:
        keys:
          type: object
          additionalProperties: true
        options:
          type: object
          properties:
            name:
              type: string
            unique:
              type: boolean
            sparse:
              type: boolean
            background:
              type: boolean
          additionalProperties: true
      required: [keys]
    ExplainResult:
      type: object
      properties:
        queryPlanner:
          type: object
          additionalProperties: true
        executionStats:
          type: object
          additionalProperties: true
        executionTimeMs:
          type: number
        indexesUsed:
          type: array
          items:
            type: string
        collectionScans:
          type: number
        totalDocsExamined:
          type: number
        totalKeysExamined:
          type: number
      required: [queryPlanner, executionTimeMs]
    IndexSuggestions:
      type: object
      properties:
        suggestions:
          type: array
          items:
            type: object
            properties:
              fields:
                type: array
                items:
                  type: string
              reason:
                type: string
              estimatedImpact:
                type: string
              createCommand:
                type: string
            required: [fields, reason, estimatedImpact, createCommand]
      required: [suggestions]
    ExportBody:
      type: object
      properties:
        format:
          type: string
          enum: [json, csv]
        filter:
          type: object
          additionalProperties: true
        projection:
          type: object
          additionalProperties: true
        limit:
          type: number
      required: [format]
    ExportResult:
      type: object
      properties:
        data:
          type: string
        format:
          type: string
        documentCount:
          type: number
        filename:
          type: string
      required: [data, format, documentCount, filename]
    ImportBody:
      type: object
      properties:
        format:
          type: string
          enum: [json, csv]
        data:
          type: string
        options:
          type: object
          properties:
            upsert:
              type: boolean
            dropFirst:
              type: boolean
          additionalProperties: true
      required: [format, data]
    ImportResult:
      type: object
      properties:
        insertedCount:
          type: number
        updatedCount:
          type: number
        errors:
          type: array
          items:
            type: string
        success:
          type: boolean
      required: [insertedCount, success]
...