# lib/api-zod/src/generated/types/aggregateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MongoVision API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AggregateBodyPipelineItem } from "./aggregateBodyPipelineItem";

export interface AggregateBody {
  pipeline: AggregateBodyPipelineItem[];
}