# lib/api-zod/src/generated/types/importBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MongoVision API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ImportBodyFormat } from "./importBodyFormat";
import type { ImportBodyOptions } from "./importBodyOptions";

export interface ImportBody {
  format: ImportBodyFormat;
  data: string;
  options?: ImportBodyOptions;
}